# backend/requirement/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: AI Task Dashboard API
  description: >
    Automatically derived CRUD API documentation based on `Data requirements.md`.
    All responses follow the standardized API response envelope.
  version: 1.1.0
servers:
  - url: /api/v1
    description: Main API Server
...
paths:
...
  /comments:
    get:
      tags: [Comments]
      summary: Get all comments
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of comments including derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/CommentResponse'
    post:
      tags: [Comments]
      summary: Create a comment
      description: >
        Creates a new comment. The `type` field is derived from the virtualId prefix
        (CH- = Challenge, ID- = Idea) and maps to Comment Types ["CH", "ID"].
        Auto-subscribes the commenter to the parent entity.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBase'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comment:
                            $ref: '#/components/schemas/CommentResponse'
...
  /comments/{id}:
    get:
      tags: [Comments]
      summary: Get comment by MongoDB _id
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '200':
          description: Comment object with derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comment:
                            $ref: '#/components/schemas/CommentResponse'
    put:
      tags: [Comments]
      summary: Update a comment
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBase'
      responses:
        '200':
          description: Comment updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comment:
                            $ref: '#/components/schemas/CommentResponse'
    delete:
      tags: [Comments]
      summary: Delete a comment
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '204':
          description: Comment deleted
  /comments/challenge/{virtualId}:
    get:
      tags: [Comments]
      summary: Get comments for a challenge by virtualId
      operationId: getCommentsByChallengeVirtualId
      description: >
        Fetches all comments for a challenge. The virtualId prefix (CH-) maps to
        Comment Type "CH". Internally resolves the challenge by virtualId and
        queries comments where type="CH" and typeId matches.
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge (e.g., CH-001)
      responses:
        '200':
          description: List of comments for the challenge
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/challenge/{virtualId}/count:
    get:
      tags: [Comments]
      summary: Get comment count for a challenge by virtualId
      operationId: getCommentCountByChallengeVirtualId
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge (e.g., CH-001)
      responses:
        '200':
          description: Comment count for the challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/idea/{virtualId}:
    get:
      tags: [Comments]
      summary: Get comments for an idea by virtualId
      operationId: getCommentsByIdeaVirtualId
      description: >
        Fetches all comments for an idea. The virtualId prefix (ID-) maps to
        Comment Type "ID". Internally resolves the idea by virtualId and
        queries comments where type="ID" and typeId matches.
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the idea (e.g., ID-0001)
      responses:
        '200':
          description: List of comments for the idea
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/idea/{virtualId}/count:
    get:
      tags: [Comments]
      summary: Get comment count for an idea by virtualId
      operationId: getCommentCountByIdeaVirtualId
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the idea (e.g., ID-0001)
      responses:
        '200':
          description: Comment count for the idea
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: Idea not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/user/{userId}:
    get:
      tags: [Comments]
      summary: Get comments by user ID
      operationId: getCommentsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: List of comments by the specified user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/CommentResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/user/{userId}/count:
    get:
      tags: [Comments]
      summary: Get comment count for a user
      operationId: getCommentCountByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID
      responses:
        '200':
          description: Comment count for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'

  # ==============================
  # ACTIVITIES ENDPOINTS
  # ==============================
...
components:
  parameters:
    MongoIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 24-character hexadecimal MongoDB Object ID
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: Amount of records to return for pagination
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
      description: Amount of records to skip for pagination
  schemas:
...
    ApiResponseEnvelope:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: string
          enum: [success, failed]
          description: Indicates whether the request was successful or failed
        message:
          type: string
          description: Human-readable success or failure message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 date-time of the response
        data:
          type: object
          nullable: true
          description: Domain-specific payload. May be null on failure.
    CountApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                count:
                  type: integer
                  description: The total count

    # ------------------
    # MINIMAL SCHEMAS (FOR REUSE)
    # ------------------
    UserMinimal:
      type: object
      required: [_id, name]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        name:
          type: string
        email:
          type: string
        companyTechRole:
          type: string
        role:
          type: string
          enum: ['ADMIN', 'MEMBER', 'USER']
...
    CommentBase:
      type: object
      properties:
        userId:
          type: string
        comment:
          type: string
        type:
          type: string
          enum: ['CH', 'ID']
        typeId:
          type: string
        createdat:
          type: string
          format: date-time
...
    CommentResponse:
      allOf:
        - $ref: '#/components/schemas/CommentBase'
        - type: object
          properties:
            _id:
              type: string
            userDetails:
              $ref: '#/components/schemas/UserMinimal'
...